修复摸鱼列表界面多次加载更多的问题。

+++ app/src/main/java/cn/cqautotest/sunnybeach/http/request/api/FishPondApi.kt
@@ -22,7 +22,7 @@ interface FishPondApi {
     @GET("${SUNNY_BEACH_BASE_URL}ct/moyu/list/{topicId}/{page}")
     suspend fun loadFishPondListById(
         @Path("topicId") topicId: String,
-        @Path("page") page: Long
+        @Path("page") page: Int
     ): BaseResponse<Fish>
 
     /**
